requirements.txt
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.10.4
certifi==2025.8.3
charset-normalizer==3.4.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
python-telegram-bot==21.6
pytz==2025.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
